requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
astroid==4.0.1
backports.asyncio.runner==1.2.0
bandit==1.8.6
black==25.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
confection==0.1.5
coverage==7.11.0
cryptography==46.0.3
cymem==2.0.11
dill==0.4.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.3.0
fastapi==0.121.0
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
murmurhash==1.0.13
mypy_extensions==1.1.0
numpy==2.2.6
packaging==25.0
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
preshed==3.0.10
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyflakes==3.4.0
Pygments==2.19.2
pylint==4.0.2
pypdf==6.1.3
pypdfium2==5.0.0
pytest==8.4.2
pytest-asyncio==1.2.0
python-docx==1.2.0
python-multipart==0.0.20
pytokens==0.2.0
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
shellingham==1.5.4
smart_open==7.4.4
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.49.3
stevedore==5.5.0
thinc==8.3.6
tomli==2.3.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==2.0.0
